# AuthorizationService/openapi_auth.yaml
openapi: 3.0.0
info:
  title: AuthorizationService API
  version: 1.0.0
  description: API для регистрации, авторизации
servers:
  - url: http://localhost:8002
    description: Локальный сервер разработки
paths:
...
  /password/change:
    put:
      summary: Изменение пароля
      description: Изменение пароля пользователя или организатора
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  description: Логин пользователя или организатора
                password:
                  type: string
                  description: Новый пароль
              required:
                - login
                - password
      responses:
        '200':
          description: Пароль успешно изменен
        '400':
          description: Неверный запрос или пользователь не найден